requirements.txt
djangorestframework
django-cors-headers
django-ckeditor-5
django-environ
django-cleanup
python-dotenv
djoser
django-db-logger
celery
redis
psycopg2
requests
pytils
django-colorfield

aiomysql
asyncpg

mysqlclient
mysql-connector-python
beautifulsoup4